requirements.txt
Django==1.5.2
djangorestframework==2.3.7
django-enumfield==1.0c1
django-allauth==0.12.0
django-celery==3.0.21
jsonfield==0.9.19
docker-py==0.2.1
GitPython==0.3.2.RC1
watchdog==0.6.0
PyYAML==3.10
psycopg2==2.5.1
south==0.8.2
pyrax==1.4.11
lockfile==0.9.1
Fabric==1.7.0
Jinja2==2.7
python-gnupg==0.3.5
django-uuidfield==0.4.0
etcd-py==0.0.5
argparse==1.2.1
gevent==0.13.8

flask-restful==0.2.5

Twisted==13.1.0
treq==0.2.0

django-nose==1.2
nose==1.3.0
coverage==3.6
testtools==0.9.32
Flask-Testing==0.4
mongomock==1.0.0
